Sketching jokes storage service and refactoring types
