+++ src/app/services/chuck-api.service.ts
@@ -1,17 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/internal/Observable';
-
-export type Joke = {
-  categories?: string[];
-  created_at?: string;
-  icon_url?: string;
-  id?: string;
-  updated_at?: string;
-  url?: string;
-  value?: string;
-  timestamp?: number;
-}
+import { Joke } from './types/joke.type';
 
 @Injectable({
   providedIn: 'root'
