+++ src/app/services/types/joke.type.ts
@@ -1,0 +1,10 @@
+export type Joke = {
+    categories?: string[];
+    created_at?: string;
+    icon_url?: string;
+    id?: string;
+    updated_at?: string;
+    url?: string;
+    value?: string;
+    timestamp?: number;
+}
